added edit category

+++ src/app/pages/category/category.component.ts
@@ -3,6 +3,7 @@ import {Component, ElementRef, OnInit, ViewChild} from '@angular/core';
 import {CategoriesService} from "../../services/categories.service";
 import {Observable} from "rxjs";
 import {ICategories} from "../../interfaces/iCategories";
+import {Router} from "@angular/router";
 
 @Component({
   selector: 'app-category',
@@ -17,23 +18,32 @@ export class CategoryComponent implements OnInit {
 
   constructor(
     private _categoriesService: CategoriesService,
+    private _router: Router,
   ) { }
 
   ngOnInit(): void {
     this.categories$ = this._categoriesService.getCategory();
   }
 
+  onCreate() {
+    this._router.navigate(['dashboard', 'create-category']).then()
+  }
 
   onDelete() {
     const inputs = this.list.nativeElement.querySelectorAll('input')
     const checkedInputs = [...inputs].filter((x: any) => x.checked)
     const categoriesIds = checkedInputs.map((x: any) => +x.id)
     if (categoriesIds.length) {
+      if (confirm('Are you sure to delete?'))
       this._categoriesService.deleteAll(categoriesIds).subscribe(
         () => {
           this.categories$ = this._categoriesService.getCategory();
         }
       )
     }
+  }
+
+  updateCategoryId(id: number) {
+    this._router.navigate(['/dashboard', 'update-category', id]).then()
   }
 }
